+++ src/components/Settings/FileNavigatorSettings/Items/HomeDirSetting.tsx
@@ -1,6 +1,6 @@
-import { useEffect, useState } from "react";
+import { useCallback, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { IconButton, ListItem, ListItemIcon, ListItemText, TextField } from "@mui/material";
+import { Box, IconButton, ListItem, ListItemIcon, ListItemText, TextField } from "@mui/material";
 import { Folder, HomeOutlined } from "@mui/icons-material";
 import { open } from "@tauri-apps/plugin-dialog";
 import { homeDir } from "@tauri-apps/api/path";
@@ -38,9 +38,24 @@ export default function HomeDirSetting() {
     initView();
   }, []);
 
-  const handleHomeDirPathChanged = async (e: React.ChangeEvent<HTMLInputElement>) => {
-    setHomeDirPath(e.target.value);
-    await settingsStore.set("home-directory", e.target.value);
+  const formAction = useCallback(
+    async (formData: FormData) => {
+      const inputPath = formData.get("path")?.toString();
+
+      if (inputPath && inputPath !== homeDirPath) {
+        setHomeDirPath(inputPath);
+        await settingsStore.set("home-directory", inputPath);
+      }
+    },
+    [homeDirPath],
+  );
+
+  const handleBlur = useCallback((e: React.FocusEvent<HTMLInputElement>) => {
+    e.currentTarget.form?.requestSubmit();
+  }, []);
+
+  const handleContextMenu = (e: React.MouseEvent<HTMLElement>) => {
+    e.stopPropagation();
   };
 
   return (
@@ -49,17 +64,22 @@ export default function HomeDirSetting() {
         <HomeOutlined />
       </ListItemIcon>
       <ListItemText primary={t("settings.file-navigator.home-directory-title")} />
-      <TextField
-        variant="standard"
-        value={homeDirPath}
-        onChange={handleHomeDirPathChanged}
-        size="small"
-        sx={{
-          width: "80%",
-          marginLeft: "4px",
-          marginRight: "4px",
-        }}
-      />
+      <Box component="form" action={formAction} sx={{ flexGrow: 1 }}>
+        <TextField
+          variant="standard"
+          key={homeDirPath}
+          name="path"
+          defaultValue={homeDirPath}
+          onContextMenu={handleContextMenu}
+          onBlur={handleBlur}
+          size="small"
+          sx={{
+            width: "80%",
+            marginLeft: "4px",
+            marginRight: "4px",
+          }}
+        />
+      </Box>
       <IconButton size="small" onClick={handleFolderClicked}>
         <Folder />
       </IconButton>